Get Sherpa screen output into HTML.
Some of Sherpa's screen output is via the Python logging instance, which
does not work well with the ipython directive. I have copied over the
IPython Sphinx extension and added in support so that it will include the
logging output. Ideally this should move upstream, but I am not yet
convinced that this is generally useful (as implemented), or is the best
way to do it.