Experimental support in UI for XSpec cflux convolution model.
This commit adds Python classes and routines that allow a user to
use the XSpec cflux convolution model from the high-level UI. This
is modelled on the existing load_conv/load_psf routines:
ConvolutionKernel and ConvolutionModel classes are added, but users
create an instance with a separate routine (in this case load_xscflux)
rather than the "normal" interface (e.g. ui.xsapec.mdl).

An example, applying the cflux model to the some of XSpec APEC and
a gaussian, is done by:

  from sherpa.astro import ui
  from sherpa.astro import xspec
  xspec.load_xscflux('cf')
  ui.set_source(cf(ui.xsapec.mdl + ui.gauss1d.line))
  cf.emin = 1
  cf.emax = 7