Add 12 XSpec convolution models.
Adds in support for the gsmooth, ireflect, kdblur, kdblur2, kerrconv,
lsmooth, partcov, rdblur, reflect, simpl, zashift, and zmshift
convolution models. Note that there is no guarantee that the results
will be sensible in Sherpa: this is primarily due to Sherpa's current
approach to evaluating models, which limits the model evaluation to the
minimum PI range (taking account of the RMF) that matches the requested
noticed range. That is, there is currently no way to extend the analysis
grid for one or more components, which is likely required to take
advantage of some of these convolution models (i.e. they need to be
evaluated on a larger grid which is then filtered down to the desired
energy range).