Export the load_xs<conv models> from ui module.
With this, the XSpec load_xs<conv model> routines are exported from
sherpa.astro.ui. Unlike the other exports from the module, this one
is dynamic, since I would rather not hard-code the names (to make it
easier to support different versions of the XSpec library, which may
add, remove, or rename models).

The test was updated to use the UI version of the function rather than
the XSpec module. It also required a change to the test to access the
model component via ui.get_model_component, as it is not available
when the test is run via 'python setup.py test', but is when the test
is run directly (e.g. 'python sherpa/astro/xspec/tests/test_xspec.py').
This is not related to the code change here, as it has been seen in
other tests of models. Given the way that the model components are
added to the global namespace, it is not surprising that there may be
differences, but I think it's something we should understand (outside
of this set of changes).