Tweak test of XSpec convolution model
This changes the test so that it uses an ARF and RMF (to check that it
works in such a situation). I am not convinced this is a great idea,
since it means the test is only run when the data files are present
and an I/O backend is available.