Get, set, and clear XFLT keywords.
Provide routines to read, write, and clear the XFLT keyword settings
used by Xspec models. Currently the models that use this are ones
that are not supported by Sherpa: smaug and the mixing models. This
is a move towards supporting them, but a method of setting the Xspec
spectrum number when evaluating one of these models needs to be added.

Note that not all of these routines are accessible using the xsFortran
API (at least in Xspec version 12.8.2). It appears impossible to set
or clear the values from this API, which means that it is going to be
hard to support this functionality in Sherpa without changing the build
requirements to require the source-build location for the include files.