Get, set, and clear the Xspec database.
Provide routines to read, write, and clear the Xspec database settings;
these are string keyword with doubel values, introduced relatively recently
to Xspec to support information flow from model to user. It is similar
to the XFLT support, except that this time it is only possible to set
a single keyword at a time, rather than input a dictionary. This makes
more sense given how the database is expected to work.

As with the XFLT keywords, most of this functionality is not available
from xsFortran (only setting a keyword to a value, so no read, query,
or clear).