Test version adding HTML output for models in IPython notebooks
This uses the code from https://bitbucket.org/doug_burke/ciao-notebooks/raw/917f54ecc759e94445b5f5b92b9dcbf66872d911/HTML%20output%20(nicer%20views).ipynb
to add HTML output for Sherpa models. Compare from an IPython session
and in an IPython notebook

    from sherpa.astro import ui
    ui.set_source(ui.polynom1d.bob + ui.gauss1d.fred)
    print(ui.get_source())
    ui.get_source()

In the IPython notebook the output of the last command is displayed
as an HTML table; in the IPython session you get the standard
repr output - e.g.

    <BinaryOpModel model instance '(polynom1d.bob + gauss1d.fred)'>

This is intended purely as a demonstration and is not intended for
production use.