sample_flux: allow parameter values at limits
This changes the limits from lo<x<hi to lo<=x<=hi. It doesn't change
the tests since they are written assuming all values are included
in the calculation (this may need to be relaxed with the change in
this commit, as it was nH=0 boundary limits that lead to #751).