WIP add basic cache tracking
This is for testing / development only

This stores basic hit/miss information on the cache, and provides
a method to display the current cache status.

This does not give reliable results when using the multi-threading
features of Sherpa. Use numcores=1 or set the parallel=False option
to avoid this.

Note that changing <mdl>.cache does not change the _queue size at
that time; this is only used when the model is next passed through
a setup call (so fit or error analysis). Perhaps it should change
the _queue at the same time?