calc_sample_flux: internal core re-arrangement
Re-order the code, so that parts are not done within the try/finally
block (as these are not going to create any logging output). It also
means that the subroutine can be expanded out directly, which makes
it a bit clearer what is going on.

Added an assert to check that the number of thawed parameters in the
fit matches the samples array. This is a pre-condition for this
routine, so it is not made into a user-friendly error.

The filtering code now uses more-native NumPy routines.