calc_sample_flux: rework and remove logging level changes
The need for changing the log level was driven by the use of the
set_source/set_full_model methods from the session class, since they
could cause a warning message. As we no-longer use these methods,
we don't need to change the log level. This also lets the try/finally
block to cover a smaller chunk of code, as it now deals with just
resetting the model parameters.

There is technically the issue that the calc_energy_flux or
setting the parameters from the samples array could lead to an
error message, but we have already run through this with the
sample_energy_flux call that is used to feed the input to this
routine (i.e. uses would have already seen something since the
sample_energy_flux routine doesn't change the log level). There
is one remaining call that *could* cause a logging message to be
displayed, namely calc_stat, but again I can not see this happening
without a message already having been seen.