Tests: test_stats rework of the tests
Prior to this commit the models being fit in the tests was not
sensible since

a) in most cases it was xsphabs + powlaw1d, which is just not meaningful
b) the response was not included, so the models were being fit in
   channel space, and so they were completely meaningless

We now apply a response to each model, and ensure a "sensible"
energy range for each fit.

Since the nH values can be a bit small, some fits are done using the
log of the nh value. Since I changed from 0.5-7 to 0.3-7 keV here
it turns out this is probably not needed (as nH values seem to be
constrained when including the bins below 0.5 keV) I have left them
in.

There is no validation that these fits are meaningful, that is this
is just a regression test and I expect tolerances will need to be
tweaked. Fortunately you can see that the best-fit parameter values
seem to converge to the same area, so at least the fits give similar
results.

The tolerances have been adjusted until the Travis tests pass.