Fix #959 - use integrate setting from the model
The integrate setting is used to define the digest value, but is
taken from the kwargs integrate flag, with a default of False.
The way the decorator is applied means that it is applied to
the calc() method, which does not have the integrate kwarg set,
rather than the method that calc calls.

There is the question of what to do if a kwarg integrate setting
is found and is different to the models integrate setting. Hopefully
this should not happen, but I wonder about some of the changes we
are making for the regrid model method (in particular in a
BinaryOpModel).