Allow model expressions to be deconstructed
The idea is to convert an aribtrary model expression, such as

  m1 * (m2 + m3)

into it's component "additive" parts, so in this case

  m1 * m2
  m1 * m3

The intention is to use this in plot_model_component when given
a model argument of None, which will then use the source expression.
In this case we want to expand out the various components, hence
these routines.

The approach for the "model expansion" part is to iterate through
the components of a model expression, expanding the parts and
then expanding the results before combining. The main part is
when combining "multiplicative" expressions (that is a
BinaryOpModel with op being *, /, or the other division types),
where we want to convert various expressions, such as
(here + stands for plus or subtract):

  (l1 + l2) * r -> l1 * r + l2 * r
  l * (r1 + r2) -> l * r1 + k * r2

  [these two should be subsumed the two above but
   currently are handled separately]
  (l1 + l2) * (r1 * r2) -> l1 * r1 * r2 + l2 * r1 * r2
  (l1 * l2) * (r1 + r2) -> l1 * l2 * r1 + l1 * l2 * r2

  (l1 + l2) * (r1 + r2) -> l1 * r1 + l2 * r1 + l1 * r2 + l2 * r2

where the trick is that any 'a * b' on the rhs should be read
as expand(expand a * expand b).

The separation of a model expression just loops through the
expression and whenever you have an "additive" BinaryOpModel
we separate out the two sides. Note that this stops at the
first non-additive-binop component, so relies on the user
having called expand (if they want them properly separated).

Since we need to create new Unary/BinaryOpModel expressions the
opstr value is added to these models to carry around. It may be
better to add some form of an "apply" method to these classes
instead (or as well as).

The current approach is to have lal the logic in functions, but
it should probably be moved into the classes themselves. This is
for a later revision.

Tests also need to be added.