Add plot_model/source_components
Add specialised routines that will display models for the
individual components of a source (or model) expression. I
had thought to make this conditional on the model argument to
plot_model/source_component but it turns out to be surprisingly
awkward, due to how we have set up these routines. I therefore
created separate routines, which is actually a little-bit
nicer.