Tests: add basic convolution-style expressions
Test out the PSF, ARF, and RMF combinations.

As we don't make it easy to combine ARF and RMF with no PHA,
I add a HackResponse1D class to do this (to mirror ARF1D and
RMF1D). Ideally we'd add this to sherpa.astro.instrument (as I've
wanted if for other cases).