Fix: make id=None work as expected for fake_pha
Without this change calling fake_pha with an id=None would lead
to strange behavior (you wouldn't use the default dataset but
start afresh).

Most users are not likely to use id=None (that is, this is unlikely
to change the behavior).

Examples of the difference, as shown in the tests:

 - use if 1 rather than None in error messages when  identifying
   the dataset
 - if there's a RMF mis-match then using id=None used to hide this
   (since it effectively wiped out the PHA dataset in use) but
   we now error out
 - the background dataset wasn't being picked up (because the
   original PHA file was being over-written)