Add a CI run using the released CIAO
Use the Linux Python 3.8 conda build of CIAO to build and test
Sherpa. This is an experiment, so all support for macOS has been
removed, and we don't need to test out various options and
package versions like we do with the existing CI runs.

We also try and build with a "more modern" setup, at least as
much as our current setup.py will let us.

The conda setup is meant to be cached. We really should also
cache the pip installation to, but that is significantly smaller
(and quicker to download).