PHA data: use explicit plot_norm flag for normalising by bin width
Rather than have the logic to determine whether a PHA dataset is
plotted as normalised-by-bin-width or not to be

  grouped or rate

add an explicit flag for it (plot_norm).

A new UI function, set_plot_opt, has been added to allow a user to
change the plot options (the new norm parameter as well as existing
factor and type arguments from set_analysis) without having to
change the analysis setting.

The aim would be to remove the plot settings from set_analysis, as
I don't think they fit there, but that's not done here.