Tests: add check of multi-plot log handling
I thought we already had a test of the situation

    set_xlog('data')
    plot_fit_resid()

to ensure that the resid plot is drawn with xlog, but apparently
not.

Note that for the background variants it's currently unclear what
the logic is of using the data/resid/bkg/bkgresid names when calling
set_xlog. We just test the current behavior.

A test of the background handling is added as doing the above made
me want to check that we understand how the "standard" set_*log
options, like "data", "resid", interact with background plots.