Clean up handling of plot_fit_xxx and xlog scaling
I had left in a note about the code being excessively complex and
it could be cleaned up, so take time to do this (now we have some
explicit checks of this code).

The plot_bkg_fit_resid handling is now a bit-more consistent, in
that both

    set_xlog('data')
    set_xlog('resid')

now do not change the plot. It is arguable that they should change
it, but that's a larger change (the previous code did not handle
resid vs bkgresid and data vs bkg here sensible I think).