Add a get_prefs method to plot objects
The idea is to allow a plot object to be able to return the
preferences without knowing the plot type. It's an open question
whether this is useful or not.

The current implementation just adds get_prefs calls to the base
classes. I did experiment with creating a base "preference" class
but this caused problems with SplitPlot. It may have been better
to use multiple inheritance via a Preferences class instead.