XSPEC: move from xsFortran to FunctionUtility interface
Rather than using the xsFortran routines - like FGSOLR - we now
directly use the FunctionUtility C++ interface. Note that the
code is based on the xsFortran.cxx file as of HEASOFT 6.28.

One additional change has been the removal of a lot of try/except
handling since these routines should not be throwing errors. It
is possible that the handler in _sherpa_init_xspec_library could
just be around FNINIT now but I have not made that change.

Unfortunately this only builds nicely against HEASOFT 6.29
(XSPEC 12.12.0) as they have normalized the header files in the
install directory. For earlier versions you have to change the
xspec_include_dirs setting of the xspec config block to

  xspec_include_dir = <build>/Xspec/src <build>/Xspec/src/include

where <build> is the path to the build (i.e. it will end in
headas-x.yy/).