Simplify group filtering in DataPHA code
The group method and _dynamic_group methods were doing similar
thing which is wasteful (but it's not a huge time sink so this is
not a time optimisation). We now do the "restore the filter"
logic only in the group call and not also in _dynamic_group.
This required tweaking things a little bit.

This code was originally written before #1219 landed and included
the comment:

  It wil be interesting to see if some of the tweaks are still
  needed once #1219 lands (where the get_filter expression
  no-longer depends on whether a dataset is grouped or not).

It would be good to track this down!