Ensure the grouping/quality column has the correct size
The grouping and quality columns must match the size of the
independent axis (the channel column here). We now require
this when ever either column is set.

Note that this is only a small part of the data checks we should
make. For instance, if we change the channel array then we should
really invalidate all the settings. This is not addressed here.