Simplify group filtering in DataPHA code
The group method and _dynamic_group methods were doing similar
things which is wasteful (but it's not a huge time sink so this
is not a time optimisation). We now do the "restore the filter"
logic only in the group call and not also in _dynamic_group.
This required tweaking things a little bit.