data: enforce areascal/backscal sizes
The values can be

 - None
 - scalar
 - array that matches the channel size (if set)

There is the possibility that we could be sent a singleton array,
that is it has a single element in, which we could treat as a
scalar (since umPy's broadcasting rules do).

The code contains commented-out code that enforced values being
positive (i.e. replacing an invalid value with 1.0). There are
existing checks in the code that do this, so the idea was to apply
it here instead, but

 - do we allow 0 values
 - a number of tests fail if we make this change

The test failures could be addressed, as they are really just
checking the current behavior (so it should be okay to change),
but I have decided to skip this step at present.