Tests: extend the no-error data tests
The data fixture is parameterized so the type of object to
be tested (e.g. Data1D or Data2DInt) can be sent in. However,
the data_no_errors fixture did not do this, and would only test
the Data class.

Fortunately we can easily remedy this, by using the data in the
POS_STATERR_ARRAY to work out what arguments to send to the
data object excluding the error values.

This means that we no test more data classes, and ones that the
user is going to use.

HOWEVER, it does mean that we no-longer run the tests with the
Data class. I think this is okay given the fact that the Data
class is not really intended for the Session case. I believe
that the test code changed here actually hid a subtle error,
as it used

    Data(name, indep, dep, ...)

rather than

    Data(name, (indep,), dep, ...)

which it should have done: note that DATA_ARGS uses (X_ARRAY,)
for the independent axis whereas before this change that
DATA_NO_ERRORS_ARGS just used X_ARRAY. Fixing that would have
meant that the test_load_arrays_no_errors test would have failed,
in the same wat that test_load_arrays_data fails. This way we
run this test wioth "valid" data types.