Tests: replace xfail data tests with passing ones
A number of tests were added, dueing #614 when the data classes
were refactored, that have been marked as xfail (presumably
because they used to work but no-longer do). These tests are
generally of the form "does the object have this method", and
they no-longer do. Since the removal of the methods was part of
the intent of the data-class refactoring, I have changed the
tests to explicitly check that the methods do not exist, hence
changing an xfail to a pass.