remove unused option linecolor
Remove `ratioline` and `xaxis` from plot.

The options have very special use - both together are used only three times in the Sherpa codebase.
They draw a horizontal line. So instead of making them an option to plot
the plot function in the calling plots can just draw a horizontal line
directly. That makes for a cleaner API everywhere.
Users should not see this change, unless they manually changed the preferences for
those lines. In this commit,
I've taken away the possibility to modify a preference to NOT draw those
lines, but I don't see any use case for that.
I think the fact that we have such a orefernece is more an
artifact of how it was implemented (every paraemters automatically
becomes an entry i nthe XXX_prefs dictionary) then a real decision to have that
as a preferences.

closes #1358