ui: rationalize the plot-preference access from Session code
The plot preferences use different names depending on whether it
is an integrated/histogram or non-integrated/line style plot, so
add a wraper routine to access this. Note that there are some
plot classes - for instance the astro ComponentSourcePlot class -
that has both line and histogram plot styles, so we have to
decide which to select. For now the histogram preferences are
selected (which better matches the existing code).