ui: check that set_xlog/... is sent a string
There are two parts:

1. check that the argument is a string, just to avoid a lower-level
   error

2. add a bunch of tests to see what names are supported by the
   set_xlog family of commands

The second point is the aim of this commit, but it turned out there
was a small improvement we could add to the code.

For the set of valid/invalid names: the code does not currently
define what is and is nto meant to be a valid name. It just happens.
As I plan to change how this works, I want to make sure we have tests
to see what may change.