ui: use common routine for string check
Use an explicit check for an argument being a string, rather than
the _check_type call. There is no functional change, it just makes
the code read slightly better.

It turns out one of the checks uses the argument as the "name"
of the field in the error message, which ismore-obvious with the
new function arguments - as it is '(meth, meth)' rather than
'(meth, "meth")' - but this is not changed here.