ui: improve internal documentation of plot/contour/image structures
Try to improve the documentation around the _plot/contour_types and
_plot/contour_type_names dictionaries. This includes making it
explicit the relationship between the two.

The relationship is simple for the sherpa/ui code, but it is
more-complex for the astro code.

Add a comment to note that the "astrocompmodel" keyword maps to
"model_componentl" and not "model_component". This typo was added
in #906.