ui: consolidate contour handling
In #1177 access to image objects from the Session class was changed
from direct access (e.g. self._xxx) to using a dictionary
(self._image_types). This commit does the same thing for the
contour objects: the objects are accessed from the _contour_types
dictionary rather than having a set of _datacontour/... fields.

For most users this will make no difference. It is only an issue
for code that directly accesses the Session class. However, these
are labelled private fields so they should be aware that things
could change.