ui: use more-descriptive names for plot types for set_xlog/...
Rather than use short names like "compsource", use the full name,
like "source_component" for the plot types. This is used by
set_xlog/... to select which plot objects get changed. In order
to support existing code aliases are provided for the old names.
So calling

    set_xlog("bkgfit")

will work, but display the message

    WARNING: The argument 'bkgfit' is deprecated and 'bkg_fit' should be used instead

This is done as a warning rather than a Python deprecated message
to make sure users see it.

This does not change the set of names that can not be used as
a dataset identifier.