ui: rework the plot/contour handling
Clean up the handling of plot and contour plots (sorry for the
repeated use of plot here). The idea is to have a single source
of truth for the labels, which are the keys of the _plot_types
and _contour_types dictionaries of the Session class. This is
not completely done, because we also include the keys of the
_plot_types_alias dictionary, so that we can allow users to use
the "old" names - like "compsource" and "bkgfit" - rather than
the "new" names - like "source_component" and "bkg_fit". The aim
is that this aliasing can be removed in a future release.

At the moment the keys of the _plot/_contour_types dictionaries
are expected to have matching get_<key>_<plot|contour> methods,
but there is no explicit check on this. Should there be?

One consequence of these changes is that we now have a consistent
set of names that are used

- by the set_xlog/... family of commands;
- plot() and contour() to refer to the plot/contour to display;
- and to determine what labels are not allowed to be used as
  a dataset identifier (because of the confusion with the
  plot/contour calls if they are used).

This does mean that the set of invalid identifiers has changed,
and there are some changes to errors that are thrown.

The use of the aliases will warn the user that they should
change the name, such as:

>>> set_ylog("bkgmodel")
WARNING: The argument 'bkgmodel' is deprecated and 'bkg_model' should be used instead
>>> set_ylog("compsource")
WARNING: The argument 'compsource' is deprecated and 'source_component' should be used instead

These are standard Python logging.WARNING messages rather
than a DeprecationWarning to make sure the user sees them
in an interactive session. It does mean that the message is
repeated each time the alias is used.

The identifier changes depend on whether you are using sherpa.ui
or sherpa.astro.ui. After

>>> from sherpa.astro.ui import *

then a call to set_default_id(label) will either pass or error
out with an IdentifierErr call with the following:

    | label            | main  | this commit |
    | ---------------- | ----- | ----------- |
    | energy           | pass  | pass        |
    | compsource       | fail  | fail        |
    | source_component | fail  | fail        |
    | bkgfit           | fail  | fail        |
    | bkg_fit          | pass  | fail *      |

where I have marked with a * those that have changed behavior.
So, we now currently exclude more keywords as the aliases are
ignored as well as the "new" labels (e.g. it used to be "bkgfit"
but it is now "bkg_fit" and "bkgfit" is the alias which will be
removed at some point).

The error change occurs in calls to plot() or contour(). Before
this commit (dropping the full traceback output):

>>> from sherpa.astro.ui import *
>>> set_xlog("boggle")
sherpa.utils.err.PlotErr: Plot type 'boggle' not found in ['data', 'model', 'source', 'fit', 'resid', 'ratio', 'delchi', 'chisqr', 'psf', 'kernel', 'compsource', 'compmodel', 'order', 'energy', 'photon', 'arf', 'bkg', 'bkgmodel', 'bkgfit', 'bkgsource', 'bkgratio', 'bkgresid', 'bkgdelchi', 'bkgchisqr']
>>> plot("boggle")
sherpa.utils.err.ArgumentErr: 'boggle' is not a valid plot type

With this commit the errors are now the same (the PlotErr version),
and the list of supported plot types has changed slightly
because of the aliases:

>>> set_xlog("boggle")
sherpa.utils.err.PlotErr: Plot type 'boggle' not found in ['data', 'model', 'source', 'fit', 'resid', 'ratio', 'delchi', 'chisqr', 'psf', 'kernel', 'source_component', 'model_component', 'order', 'arf', 'bkg', 'bkg_model', 'bkg_fit', 'bkg_source', 'bkg_ratio', 'bkg_resid', 'bkg_delchi', 'bkg_chisqr']
>>> plot("boggle")
sherpa.utils.err.PlotErr: Plot type 'boggle' not found in ['data', 'model', 'source', 'fit', 'resid', 'ratio', 'delchi', 'chisqr', 'psf', 'kernel', 'source_component', 'model_component', 'order', 'arf', 'bkg', 'bkg_model', 'bkg_fit', 'bkg_source', 'bkg_ratio', 'bkg_resid', 'bkg_delchi', 'bkg_chisqr']