ui: plot object access via a function
Following image and contour handling, the plot access to
the plot objects is now handled via a function. This means
that the plot objects for most, but not all, plots are
now accessed indirectly, rather than by grabbing a named
field from the session object.

The direct-access fields are retained in this commit; they
will be removed in a later commit.

Note that plot_source_component has changed for PHA data, as
it now includes the source template class. This code stands out
from the other cases, and it needs investigating (as we currently
don't have much testing of the behavior for template models).