Tests: change plot tests to access objects via API
A number of tests had been sing low-level/direct access to the
plot objects stored in the session classes. This direct access
makes it hard to change the code, so switch the code to using
an acual API - normally the get_foo_plot call. There are issues
with making sure to call it with

- the correct id (so that we can get the type-specific behavior
  for certain tests)
- with recalc=False so we test what we want to be testing
  (which is the previous plot command)

Given that a number of tests of plot_foo can be done with just
a get_foo_plot call (with recalc=True), given how the two calls
are related, we could probably redo some of these tests, but
that is not done here.

There are some cases where we don't have a direct API to the
plot object: for PHA data sets the fit and bkg_fit plots use
a specialized class for the model (it ignores the grouping of
the PHA dataset). The only way to get at this "sensibly" is
with get_[bkg_]fit_plot(id, recalc=False).modelplot