Tests: add more tests of set_xlog/.. handling of component plots
There's some subtle issues gonig on with the source and model
component plots, so add some tests to check this out. These are
currently regression tests. The test_set_plot_opt_changes_fields
routine is acting as a "this shoule behave sensibly" style test
so is marked as xfail for the behavior under test here.