Fix tests after rebase
Part of the rebase made tests fail now, e.g. because
new tests with `requires_plotting` has been added, but that
decorator has been removed in this branch.

Also, the color selection in `OrderPlot` failed - in fact
it never produced matplotlib compatible colors. That indicates
 that the colors made there were never propagated through to the plot.
Only after the changes made in the last few commits is the color used
and thus the test fails.
So, this is essentially an unrealted bugfix, but since it needs to be
worked into the new "backends are modules" scheme
it is treated here.