Tests: add checks of ignore_bad handling
The idea is to add tests of behavior you could see with ignore_bad so
we can use these as regression tests.

I have also added a test that shows what appears to be a bug in the
group library (just reported to the CXC team). If this gets fixed (and
is not just my misunderstanding of what the adaptive grouping is
doing) then the test will need updating.  This test is called
test_pha_group_adapt_check.