fake: add a method argument
The data simulation done by fake_pha is now controllable by the
user, so it can be set to something other that sherpa.utils.poisson_noise.

Thiss primarily for testing - and a few test have been updated to
inclue this - which makes it easier to see when fake_pha is not
working correctly. It also follows the sherpa.ui.utils.fake method,
which uses the name method for this argument (but in this commit
the sherpa.astro.ui.utils.fake_pha method has not been updated to
accept the method argument).